+++ lib/src/functions/picker_functions.dart
@@ -127,24 +127,6 @@ ColorSwatch<Object?>? findColorSwatch(
     }
   }
   return (color is ColorSwatch && swatches.contains(color)) ? color : null;
-
-  // final ColorSwatch<Object?>? foundSwatch =
-  //     (color is ColorSwatch && swatches.contains(color)) ? color : null;
-  //
-  // if (foundSwatch != null) {
-  //   debugPrint('findColorSwatch $foundSwatch');
-  //   return foundSwatch;
-  // }
-  //
-  // // We did not find any matching color. We try with no opacity, as well.
-  // for (final ColorSwatch<Object> mainColor in swatches) {
-  //   if (isShadeOfMain(mainColor, color.withAlpha(0xFF), include850)) {
-  //     return mainColor;
-  //   }
-  // }
-  // return (color is ColorSwatch && swatches.contains(color.withAlpha(0xFF)))
-  //     ? color
-  //     : null;
 }
 
 /// Check if a given color is a shade of the main color, return true if it is.
@@ -185,10 +167,10 @@ List<Color> getMaterialColorShades(ColorSwatch<Object> color, bool include850) {
 }
 
 /// Return the M3 tonal palette for a passed in color as a list of Colors.
-List<Color> getTonalColors(Color color) {
+List<Color> getTonalColors(Color color, bool fixedMinChroma) {
   final Cam16 camColor = Cam16.fromInt(color.value);
-  final FlexTonalPalette tonalColors =
-      FlexTonalPalette.of(camColor.hue, math.max(48, camColor.chroma));
+  final FlexTonalPalette tonalColors = FlexTonalPalette.of(camColor.hue,
+      fixedMinChroma ? math.max(48, camColor.chroma) : camColor.chroma);
 
   // ignore: unnecessary_lambdas
   return tonalColors.asList.map((int e) => Color(e)).toList();