+++ test/copy_paste_handler_test.dart
@@ -30,6 +30,7 @@ void main() {
             onCopyPasteMenuOpened: () {},
             focusNode: focusNode,
             autoFocus: true,
+            noPasteIntent: false,
             child: Container(),
           ),
         ),